% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-old.R
\name{convert}
\alias{convert}
\title{Convert alphas to gammas}
\usage{
convert(betas.and.alphas, epsilon = 1e-05, group = group,
  main.effect.names.list, interaction.names.list)
}
\arguments{
\item{betas.and.alphas}{q x 1 data.frame or matrix of main effects and
interaction estimates. For example the output from the \code{uni_fun}
function. The rownames must be appropriately labelled because these labels
will be used in other functions}

\item{epsilon}{threshold to avoid division by a very small beta e.g. if any
of the main effects are less than epsilon, set gamma to zero. This should
not really be an important parameter because this function is only used in
the initialization step, where the intial estimates are from OLS or ridge
regression and therefor should not be very close to 0}

\item{main.effect.names}{character vector of main effects names. MUST be
ordered in the same way as the column names of \code{x}. e.g. if the column
names of \code{x} are \code{\"x1\",\"x2\"} then \code{main.effect.names =
c("x1","x2")}}

\item{interaction.names}{character vector of interaction names. MUST be
separated by a colon (e.g. x1:x2), AND MUST be
ordered in the same way as the column names of \code{x}}
}
\value{
a labelled q x 1 data.frame of betas and gammas
}
\description{
function that takes a vector of betas (which are the main
  effects) and alphas (which are the interaction effects) and converts the
  alphas to gammas.
}
\details{
note that \deqn{y = \beta_0 + \beta_1 x_1 + ... + \beta_p x_p +
  \alpha_{12} x_1 x_2 + ... + \alpha_{p-1,p} x_p x_{p-1} } and
  \deqn{\alpha_{ij} = \gamma_{ij} * \beta_i*\beta_j , i < j}

  This function is used because the fitting algorithm estimates the gammas,
  and furthermore, the L1 penalty is placed on the gammas. It is used only in
  the initialization step in the \code{\link{sail}} function
}
\seealso{
\code{\link{sail}}, \code{\link{Q_theta}}
}
