% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-old.R
\name{convert2}
\alias{convert2}
\title{Convert gammas to alphas}
\usage{
convert2(beta, gamma, main.effect.names, interaction.names, group = group,
  intercept = NULL)
}
\arguments{
\item{main.effect.names}{character vector of main effects names. MUST be
ordered in the same way as the column names of \code{x}. e.g. if the column
names of \code{x} are \code{\"x1\",\"x2\"} then \code{main.effect.names =
c("x1","x2")}}

\item{interaction.names}{character vector of interaction names. MUST be
separated by a colon (e.g. x1:x2), AND MUST be
ordered in the same way as the column names of \code{x}}

\item{betas.and.gammas}{q x 1 data.frame or matrix of betas and gamma
estimates. For example the output from the \code{convert} function. The
rownames must be appropriately labelled because these labels will be used
in other functions}
}
\value{
a labelled q x 1 data.frame of betas and alphas
}
\description{
function that takes a vector of betas (which are the main
  effects) and gammas and converts the alphas to gammas. This function is
  used to calculate the linear predictor of the likelihood function (the Q
  function in the fitting algorithm)
}
