% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createfolds}
\alias{createfolds}
\title{Create CV Folds}
\usage{
createfolds(y, k = 10, list = FALSE, returnTrain = FALSE)
}
\arguments{
\item{y}{vector of response}

\item{an}{integer for the number of folds.}
}
\value{
A vector of CV fold ID's for each observation in \code{y}
}
\description{
\code{createfolds} splits the data into \code{k} groups. Taken
  from the \code{caret} package (see references for details)
}
\details{
For numeric y, the sample is split into groups sections based on
  percentiles and sampling is done within these subgroups
}
\references{
\url{http://topepo.github.io/caret/splitting.html}
}
