% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv_lspath}
\alias{cv_lspath}
\alias{cvcompute}
\alias{getmin_type}
\alias{lambda.interp}
\title{Compute cross validation error}
\usage{
cv_lspath(outlist, y, df, foldid, design, nlambda, nlambda.beta, nlambda.gamma)

cvcompute(mat, foldid, nlams)

getmin_type(lambda, cvm, cvsd, type)

lambda.interp(lambda, s)
}
\arguments{
\item{outlist}{list of cross validated fitted models. List is of length equal
to \code{nfolds} argument in \code{\link{cv.sail}} function}

\item{y}{response variable (matrix form) of dimension \code{n x 1}}

\item{foldid}{numeric vector indicating which fold each observation belongs
to}

\item{nlambda}{total number of tuning parameter pairs. This includes those
pairs of tuning parameters that didn't converge in the CV folds.}

\item{nlambda.beta}{number of tuning paramters for the main effect}

\item{nlambda.gamma}{number of tuning parameters for the interaction effects}

\item{type}{should be one of "beta" or "gamma"}

\item{outlist}{list containing results from cv.sail function. each element of
the list is a run of sail_multiple_faster for each CV fold}
}
\description{
functions used to calculate cross validation error and used by
  the \code{\link{cv.sail}} function
}
\details{
The output of the \code{cv_lspath} function only returns values for those tuning
  paramters that DID converge
}
\section{Functions}{
\itemize{
\item \code{cvcompute}: Computations for crossvalidation error

\item \code{getmin_type}: Function that returns the tuning paramter corresponding
to the minimum cross validated error and cross validated error within 1
standard error of the minimum

\item \code{lambda.interp}: Interpolation function. Currently not implemented.
}}

\seealso{
\code{\link{cv.sail}}
}
