% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gendataPaper}
\alias{gendataPaper}
\title{Simulation Scenarion from Bhatnagar et al. (2018+) sail paper}
\usage{
gendataPaper(n, p, corr = 0, E = truncnorm::rtruncnorm(n, a = -1, b = 1),
  betaE = 2, SNR = 2, hierarchy = c("strong", "weak", "none"),
  nonlinear = TRUE, interactions = TRUE)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{p}{PARAM_DESCRIPTION}

\item{corr}{PARAM_DESCRIPTION, Default: 0}

\item{E}{PARAM_DESCRIPTION, Default: truncnorm::rtruncnorm(n, a = -1, b = 1)}

\item{betaE}{PARAM_DESCRIPTION, Default: 2}

\item{SNR}{PARAM_DESCRIPTION, Default: 2}

\item{hierarchy}{PARAM_DESCRIPTION, Default: c("strong", "weak", "none")}

\item{nonlinear}{PARAM_DESCRIPTION, Default: TRUE}

\item{interactions}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[truncnorm]{truncnorm}}
 \code{\link[stats]{rnorm}},\code{\link[stats]{cor}}
}
