% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%dopar\%}
\alias{\%dopar\%}
\title{Dopar function so that foreach stays in Imports and not Depends}
\usage{
obj \%dopar\% ex
}
\description{
Dopar function so that foreach stays in Imports and not Depends
}
