% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{lspath}
\alias{lspath}
\title{R Source code file for fitting functions
this is where most of the work is being done
not exported
Author: Sahir Bhatnagar
Created: 2016
Updated: April 6, 2018
Gaussian Response fitting function with warm starts}
\usage{
lspath(x, y, e, basis, center.x, center.e, expand, group, group.penalty,
  weights, nlambda, thresh, fdev, maxit, verbose, alpha, nobs, nvars, vp, we,
  wj, wje, flmin, vnames, ne, ulam)
}
\description{
R Source code file for fitting functions
this is where most of the work is being done
not exported
Author: Sahir Bhatnagar
Created: 2016
Updated: April 6, 2018
Gaussian Response fitting function with warm starts
}
