% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oasis}
\alias{oasis}
\title{OASIS Brain Data}
\format{A list with 3 elements: \describe{ \item{x}{a matrix of dimension
  \code{136 x 30} with the following colums:\describe{\item{Age}{Patient's
  age} \item{EDUC}{Education} \item{MMSE}{Mini-Mental State Exam}
  \item{eTIV}{Estimated Total Intracranial Volume} \item{nWBV}{Normalized
  Whole Brain Volume} \item{ASF}{Atlas scaling factor}
  \item{noise[1-24]}{24 independent standard normal noise variables}}}\item{y}{a
  numeric vector of length 136 representing the right Hippocampal volume for
  each patient}\item{e}{a binary 0/1 vector of length 136, representing
  Dementia status. 0: Non-demented, 1: Demented} }}
\source{
\url{https://github.com/stnava/RMI/tree/master/tomfletcher}

\url{http://www.oasis-brains.org/}
}
\usage{
oasis
}
\description{
A dataset containing a small subset of the OASIS Brain Data project. Noise
variables have been added to increase the number of predictors and show the
utility of the sail package.
}
\examples{
oasis
}
\keyword{datasets}
