% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv.shim}
\alias{plot.cv.shim}
\title{Plot the cross-validation curve produced by cv.shim}
\usage{
\method{plot}{cv.shim}(x)
}
\arguments{
\item{x}{fitted \code{cv.shim} object}
}
\description{
Plots the cross-validation curve, and upper and lower standard
  deviation curves, as a function of the \eqn{\lambda_\beta} and
  \eqn{\lambda_\gamma} values used. Using \code{ggplot2} facet plots, each
  facet represents a unique value for \eqn{\lambda_\gamma}, and the x-axis is
  the sequence of corresponding \eqn{\lambda_\beta}
}
\details{
A plot is produced, and nothing is returned. A colored vertical line
  is drawn at the pair of tuning parameters that leads to the minimum CV
  error and another is drawn at the 1 standard error rule pair of tuning
  parameters
}
\seealso{
\code{\link{shim}} and \code{\link{cv.shim}}
}
\author{
Sahir Bhatnagar

Maintainer: Sahir Bhatnagar \email{sahir.bhatnagar@mail.mcgill.ca}
}
