% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCoefSail}
\alias{plotCoefSail}
\title{Plot the coefficient plot produced by sail}
\usage{
plotCoefSail(beta, norm, lambda, df, dev, label = FALSE, xvar = c("norm",
  "lambda", "dev"), xlab = iname, ylab = "Coefficients", ...)
}
\description{
Plot the coefficient plot produced by sail
}
