% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCoefShim}
\alias{plotCoefShim}
\title{Plot the coefficient plot produced by shim}
\usage{
plotCoefShim(beta, norm, lambda, df, dev, label = FALSE, xvar = c("norm",
  "lambda", "dev"), xlab = iname, ylab = "Coefficients", ...)
}
\description{
Plot the coefficient plot produced by shim
}
