% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.sail}
\alias{predict.sail}
\alias{coef.sail}
\title{Make predictions from a sail object}
\usage{
\method{predict}{sail}(object, newx, newe, s = NULL, type = c("link",
  "response", "coefficients", "nonzero", "class"), ...)

\method{coef}{sail}(object, s = NULL, exact = FALSE, ...)
}
\arguments{
\item{object}{Fitted sail model object}

\item{s}{index of tuning parameter. Must be a character and an element of
"s1","s2",...."s100", where "s100" is the index of the last pair of tuning
parameters. Default is \code{NULL}}

\item{type}{Type of prediction required. Type "link" gives the fitted values.
Type "response" for "gaussian" type "response" is equivalent to type
"link". Type "coefficients" computes the coefficients at the requested
values for s. Type "nonzero" returns a list of the indices of the nonzero
coefficients for each value of s.}
}
\description{
this function only works for tuning parameter values defined by
  the sail_multiple_faster function. The interpolation feature is not working
  yet

Get coefficients from a "sail" object
}
