% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Q_theta}
\alias{Q_theta}
\title{Likelihood function}
\usage{
Q_theta(R, nobs, lambda, alpha, we, wj, wje, betaE, theta_list, gamma)
}
\arguments{
\item{gamma}{p*(p-1)/2 x 1 matrix of gamma estimates}

\item{beta}{p x 1 matrix of main effect estimates}

\item{weights}{adaptive weights calculated by \code{ridge_weights} function
with rownames corresponding to column names of x}

\item{lambda.beta}{a single tuning parameter for main effects}

\item{lambda.gamma}{a single tuning parameter for gammas}

\item{main.effect.names}{character vector of main effects names}

\item{interaction.names}{character vector of interaction names. must be
separated by a colon (e.g. \code{x1:E})}
}
\value{
value of likelihood function
}
\description{
calculates likelihood function. Used to assess convergence of
  fitting algorithm. This corresponds to the Q(theta) function in the paper
}
\note{
you dont use the intercept in the calculation of the Q function
because its not being penalized
}
