% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-old.R
\name{ridge_weights}
\alias{ridge_weights}
\title{Calculate Adaptive Weights based on Ridge Regression}
\usage{
ridge_weights(x, y, group, main.effect.names.list, interaction.names.list,
  include.intercept = F)
}
\arguments{
\item{x}{Design matrix of dimension \code{n x q}, where \code{n} is the
number of subjects and q is the total number of variables; each row is an
observation vector. This must include all main effects and interactions as
well, with column names corresponding to the names of the main effects
(e.g. \code{x1, x2, E}) and their interactions (e.g. \code{x1:E, x2:E}).
All columns should be scaled to have mean 0 and variance 1; this is done
internally by the \code{\link{sail}} function.}

\item{y}{response variable (matrix form) of dimension \code{n x 1}}

\item{include.intercept}{logical if intercept should be fitted. Default is
\code{FALSE}. Should be set to \code{TRUE} if \code{y} is not centered}

\item{main.effect.names}{character vector of main effects names}

\item{interaction.names}{character vector of interaction names. MUST be
separated by a colon (e.g. \code{x1:e, x2:e})}
}
\value{
\code{q x 1} matrix of weights for the main effects and interaction
  terms
}
\description{
uses ridge regression from \code{glmnet} package to calculate
  the adaptive weights used in the fitting algorithm implemented in the
  \code{sail} function.
}
\details{
Ridge regression is performed using the
  \code{\link[glmnet]{cv.glmnet}} function and the tuning parameter is chosen
  using 10 fold cross validation
}
\examples{
\dontrun{
# number of observations
n <- 100

# number of predictors
p <- 5

# environment variable
e <- sample(c(0,1), n, replace = T)

# main effects
x <- cbind(matrix(rnorm(n*p), ncol = p), e)

# need to label columns
main_effect_names <- c(paste0("x",1:p), "e")
interaction_names <- paste0("x",1:p, ":e")
dimnames(x)[[2]] <- main_effect_names

# design matrix without intercept
X <- model.matrix(~(x1+x2+x3+x4+x5)*e-1, data = as.data.frame(x))

# response
Y <- X \%*\% rbinom(ncol(X), 1, 0.2) + 3*rnorm(n)

# standardize data
data_std <- standardize(X,Y)

ridge_weights(x = data_std$x, y = data_std$y,
              main.effect.names = main_effect_names,
              interaction.names = interaction_names)
}
}
\references{
Friedman, J., Hastie, T. and Tibshirani, R. (2008)
  \emph{Regularization Paths for Generalized Linear Models via Coordinate
  Descent}, \url{http://www.stanford.edu/~hastie/Papers/glmnet.pdf}
  \emph{Journal of Statistical Software, Vol. 33(1), 1-22 Feb 2010}
  \url{http://www.jstatsoft.org/v33/i01/}
}
\author{
Sahir Bhatnagar

  Maintainer: Sahir Bhatnagar \email{sahir.bhatnagar@mail.mcgill.ca}
}
