% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sailsim}
\alias{sailsim}
\title{Simulated Data Used in Bhatnagar et al. (2018+) Paper}
\format{A list with 3 elements: \describe{ \item{x}{a matrix of dimension
  \code{100 x 200} where rows are observations and columns are
  predictors}\item{y}{a numeric response vector of length 100 }\item{e}{a
  numeric exposure vector of length 100} }}
\usage{
sailsim
}
\description{
A dataset containing simulated data used in the accompanying paper to this
package
}
\details{
The code used to simulate the data is available at \url{}. The
  additive functions used in the simulation are based on those in Huang et
  al. (2010). The true model is given by \deqn{Y = f1(X1) + f2(X2) +  f3(X3)
  + f4(X4) +  E * (2 +  f3(X3) +  f4(X4))} where
  \describe{\item{}{f1(t)=5t}\item{}{f2(t)=4.5(2t - 1)^2}\item{}{f3(t)=
  4sin(2pi*t) / (2-sin(2pi*t)}\item{}{f4(t)=6(0.1sin(2pi*t) + 0.2cos(2pi*t) +
  0.3sin(2pi*t)^2 + 0.4cos(2pi*t)^3 + 0.5sin(2pi*t)^3)}}
}
\references{
Huang J, Horowitz JL, Wei F. Variable selection in nonparametric
  additive models. Annals of statistics. 2010 Aug 1;38(4):2282.

Bhatnagar SR, Yang Y, Greenwood CMT. Sparse additive interaction
  models with the strong heredity property (2018+). Preprint.
}
\keyword{datasets}
