% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shim_once}
\alias{shim_once}
\title{Fit Strong Heredity model with one iteration}
\usage{
shim_once(x, y, main.effect.names, interaction.names,
  initialization.type = c("ridge", "univariate"), nlambda.gamma = 20,
  nlambda.beta = 20, lambda.factor = ifelse(nobs < nvars, 0.01, 1e-06))
}
\arguments{
\item{x}{Design matrix of dimension \code{n x q}, where \code{n} is the
number of subjects and q is the total number of variables; each row is an
observation vector. This must include all main effects and interactions as
well, with column names corresponding to the names of the main effects
(e.g. \code{x1, x2, E}) and their interactions (e.g. \code{x1:E, x2:E}).
All columns should be scaled to have mean 0 and variance 1; this is done
internally by the \code{\link{shim}} function.}

\item{y}{response variable (matrix form) of dimension \code{n x 1}}

\item{main.effect.names}{character vector of main effects names}

\item{interaction.names}{character vector of interaction names. MUST be
separated by a colon (e.g. \code{x1:e, x2:e})}

\item{initialization.type}{The procedure used to estimate the regression
coefficients and used in the \code{\link{uni_fun}} function. If
\code{"univariate"} then a series of univariate regressions is performed
with the response variable \code{y}. If \code{"ridge"} then ridge
regression is performed using the \code{\link[glmnet]{cv.glmnet}} function
and the tuning parameter is chosen using 10 fold cross validation. The
default is \code{"ridge"}.}

\item{nlambda.gamma}{number of tuning parameters for gamma}

\item{nlambda.beta}{number of tuning parameters for beta}

\item{lambda.factor}{The factor for getting the minimal lambda in lambda
sequence, where \code{min(lambda) = lambda.factor * max(lambda).
max(lambda)} is the smallest value of lambda for which all coefficients are
zero. The default depends on the relationship between \code{N} (the number
of rows in the matrix of predictors) and \code{p} (the number of
predictors). If \code{N > p}, the default is \code{1e-6}, close to zero. If
\code{N < p}, the default is \code{0.01}. A very small value of
lambda.factor will lead to a saturated fit.}
}
\value{
list of length 2, first element is \code{lambda_gamma} and second
  element is \code{lambda_beta}
}
\description{
This function runs the first iteration of the fitting algorithm
  just to get the sequence of \code{lambda_gamma} and \code{lambda_beta}
}
\details{
A unique sequence of tuning parameters for the main effects
  (lambda.beta) is calculated for each tuning parameter for the interaction
  terms (lambda.gamma)
}
\examples{
# number of observations
n <- 100

# number of predictors
p <- 5

# environment variable
e <- sample(c(0,1), n, replace = T)

# main effects
x <- cbind(matrix(rnorm(n*p), ncol = p), e)

# need to label columns
dimnames(x)[[2]] <- c("x1","x2","x3","x4","x5","e")

# design matrix without intercept (can be user defined interactions)
X <- model.matrix(~(x1+x2+x3)*e+x1*x4+x3*x5-1, data = as.data.frame(x))

# names must appear in the same order as X matrix
interaction_names <- grep(":", colnames(X), value = T)
main_effect_names <- setdiff(colnames(X), interaction_names)

# response
Y <- X \%*\% rbinom(ncol(X), 1, 0.6) + 3*rnorm(n)

# standardize data
data_std <- standardize(X,Y)

shim_once(x = data_std$x, y = data_std$y,
          main.effect.names = main_effect_names,
          interaction.names = interaction_names,
          nlambda.gamma = 5, nlambda.beta = 5)
}
\seealso{
\code{\link{shim}}
}
\author{
Sahir Bhatnagar

Maintainer: Sahir Bhatnagar \email{sahir.bhatnagar@mail.mcgill.ca}
}
