% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{soft}
\alias{soft}
\title{Soft Thresholding Function}
\usage{
soft(x, y, beta, lambda, weight)
}
\arguments{
\item{x}{Design matrix of dimension \code{n x q}, where \code{n} is the
number of subjects and q is the total number of variables; each row is an
observation vector. This must include all main effects and interactions as
well, with column names corresponding to the names of the main effects
(e.g. \code{x1, x2, E}) and their interactions (e.g. \code{x1:E, x2:E}).
All columns should be scaled to have mean 0 and variance 1; this is done
internally by the \code{\link{sail}} function.}

\item{y}{response variable (matrix form) of dimension \code{n x 1}}

\item{beta}{vector of regression coefficients to be thresholded}

\item{lambda}{tuning parameters}

\item{weight}{vector of weights for each beta}
}
\value{
matrix of thresholded regression coefficients
}
\description{
Soft Thresholding Function
}
\note{
user must supply \code{x} AND \code{y}, or \code{beta}, but not both. I
  set it up this way because to get the sequence of lambdas, I use the
  \code{beta} argument so that I only compute this once. I use the \code{x,
  y} argument for the CV folds. \code{lambda} can be a vector and this
  functions will return each thresholded beta for each lambda
}
\examples{
 soft(beta = 0.5, lambda = seq(0.001,0.65,length.out = 10), weight = 1)
}
\author{
Sahir Bhatnagar

Maintainer: Sahir Bhatnagar \email{sahir.bhatnagar@mail.mcgill.ca}
}
