% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-old.R
\name{uni_fun}
\alias{uni_fun}
\title{Univariate regressions}
\usage{
uni_fun(x, y, type = c("ridge", "univariate"), variables,
  include.intercept = FALSE)
}
\arguments{
\item{x}{Design matrix of dimension \code{n x q}, where \code{n} is the
number of subjects and q is the total number of variables; each row is an
observation vector. This must include all main effects and interactions as
well, with column names corresponding to the names of the main effects
(e.g. \code{x1, x2, E}) and their interactions (e.g. \code{x1:E, x2:E}).
All columns should be scaled to have mean 0 and variance 1; this is done
internally by the \code{\link{sail}} function.}

\item{y}{response variable (matrix form) of dimension \code{n x 1}}

\item{type}{The procedure used to estimate the regression coefficients. If
\code{"univariate"} then a series of univariate regressions is performed
with the response variable \code{y}. If \code{"ridge"} then ridge
regression is performed using the \code{\link[glmnet]{cv.glmnet}} function
and the tuning parameter is chosen using 10 fold cross validation. The
default is \code{"ridge"}.}

\item{variables}{character vector of variable names for which you want the
univariate regression estimate. Must be contained in the column names of
\code{x}. This only applies if \code{type="univariate"}.}

\item{include.intercept}{Should intercept be fitted (default is
\code{FALSE}). Should be set to \code{TRUE} if \code{y} is not centered}
}
\value{
Regression coefficients as a \code{q x 1 data.frame}
}
\description{
Function used to create initial estimates in fitting algorithm
  of the strong heredity interaction model implemented in the
  \code{\link{sail}} function
}
\note{
\code{p} is defined as the number of main effects. I have introduced
  \code{q} as being the total number of variables (e.g. the number of columns
  in the design matrix).
}
\examples{
# number of observations
n <- 100

# number of predictors
p <- 5

# environment variable
e <- sample(c(0,1), n, replace = T)

# main effects
x <- cbind(matrix(rnorm(n*p), ncol = p), e)

# need to label columns
dimnames(x)[[2]] <- c(paste0("x",1:p), "e")

# design matrix without intercept
X <- model.matrix(~(x1+x2+x3+x4+x5)*e-1, data = as.data.frame(x))

# response
Y <- X \%*\% rbinom(ncol(X), 1, 0.2) + 3*rnorm(n)

uni_fun(X, Y)
}
\seealso{
\code{\link{sail}}, \code{\link[glmnet]{cv.glmnet}}
}
\author{
Sahir Bhatnagar

Maintainer: Sahir Bhatnagar \email{sahir.bhatnagar@mail.mcgill.ca}
}
