% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_weights}
\alias{update_weights}
\title{Update Weights based on betas and gammas.}
\usage{
update_weights(betas, gammas, main.effect.names, interaction.names,
  epsilon = 1e-05)
}
\arguments{
\item{main.effect.names}{character vector of main effects names}

\item{interaction.names}{character vector of interaction names. must be
separated by a colon (e.g. x1:x2)}

\item{betas.and.gammas}{q x 1 data.frame or matrix of betas and gamma
estimates. The rownames must be appropriately labelled because these labels
are be used in this function and must match those in the arguments
\code{main.effect.names} and \code{interaction.names}}
}
\value{
q x 1 matrix of weights
}
\description{
uses betas and gammas to update weights. this is used to update
  the weights at each iteration of the fitting algorithm in the \code{shim}
  function
}
