% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xtilde}
\alias{xtilde}
\title{Calculate working X's to update Gammas.}
\usage{
xtilde(interaction.names, data.main.effects, beta.main.effects)
}
\arguments{
\item{interaction.names}{character vector of interaction names. must be
separated by a ':' (e.g. x1:x2)}

\item{data.main.effects}{data frame or matrix containing the main effects
data}

\item{beta.main.effects}{data frame or matrix containing the coefficients of
main effects}

\item{nlambda}{number of tuning parameters}
}
\value{
matrix of working X's (xtilde)
}
\description{
function used to calculate working X's (xtilde) in step 3 of
  algorithm
}
