% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-old.R
\name{xtilde_mod}
\alias{xtilde_mod}
\title{Calculate working X's to update Betas}
\usage{
xtilde_mod(x, group, main.effect.names.list, interaction.names.list,
  beta.main.effects, gamma.interaction.effects)
}
\arguments{
\item{beta.main.effects}{data frame or matrix containing the coefficients of
main effects}

\item{gamma.interaction.effects}{data frame or matrix containing the gamma
parameters}

\item{interaction.names}{character vector of interaction names. must be
separated by a ':' (e.g. x1:x2)}

\item{data.main.effects}{data frame or matrix containing the main effects
data}
}
\value{
matrix of working X's (xtilde) of dimension n x (p*(p-1)/2)
}
\description{
function used to calculate working X's (xtilde) in step 4 of
  algorithm
}
\note{
this function is a modified x_tilde for step 4 because we thought maybe
  there was a typo. Math and results suggests that there is a typo in the
  original paper.
}
