% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sail.R
\name{createfolds}
\alias{createfolds}
\title{Create CV Folds}
\usage{
createfolds(y, k = 10, list = FALSE, returnTrain = FALSE)
}
\arguments{
\item{y}{vector of response}

\item{k}{integer for the number of folds.}

\item{list}{logical - should the results be in a list (TRUE) or a matrix}

\item{returnTrain}{a logical. When true, the values returned are the sample
positions corresponding to the data used during training. This argument
only works in conjunction with \code{list = TRUE}}
}
\value{
A vector of CV fold ID's for each observation in \code{y}
}
\description{
\code{createfolds} splits the data into \code{k} groups. Taken
  from the \code{caret} package (see references for details)
}
\details{
For numeric y, the sample is split into groups sections based on
  percentiles and sampling is done within these subgroups
}
\references{
\url{http://topepo.github.io/caret/splitting.html}
}
