% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-old.R, R/functions.R
\name{cv.lspath}
\alias{cv.lspath}
\alias{getmin_type}
\alias{lambda.interp}
\alias{lamfix}
\alias{cvcompute}
\alias{getmin}
\title{Compute cross validation error}
\usage{
cv.lspath(outlist, lambda, x, y, e, weights, foldid, type.measure, grouped,
  keep = FALSE)

getmin_type(lambda, cvm, cvsd, type)

lambda.interp(lambda, s)

lamfix(lam)

cvcompute(mat, weights, foldid, nlams)

getmin(lambda, cvm, cvsd)
}
\arguments{
\item{outlist}{list of cross validated fitted models. List is of length equal
to \code{nfolds} argument in \code{\link{cv.sail}} function}

\item{x}{Design matrix of dimension \code{n x q}, where \code{n} is the
number of subjects and q is the total number of variables; each row is an
observation vector. This must include all main effects and interactions as
well, with column names corresponding to the names of the main effects
(e.g. \code{x1, x2, E}) and their interactions (e.g. \code{x1:E, x2:E}).
All columns should be scaled to have mean 0 and variance 1; this is done
internally by the \code{\link{sail}} function.}

\item{y}{response variable (matrix form) of dimension \code{n x 1}}

\item{foldid}{numeric vector indicating which fold each observation belongs
to}

\item{type}{should be one of "beta" or "gamma"}

\item{nlambda.beta}{number of tuning paramters for the main effect}

\item{nlambda.gamma}{number of tuning parameters for the interaction effects}

\item{nlambda}{total number of tuning parameter pairs. This includes those
pairs of tuning parameters that didn't converge in the CV folds.}

\item{outlist}{list containing results from cv.sail function. each element of
the list is a run of sail_multiple_faster for each CV fold}
}
\description{
functions used to calculate cross validation error and used by
  the \code{\link{cv.sail}} function
}
\details{
The output of the \code{cv.lspath} function only returns values for those tuning
  paramters that DID converge
}
\section{Functions}{
\itemize{
\item \code{getmin_type}: Function that returns the tuning paramter corresponding
to the minimum cross validated error and cross validated error within 1
standard error of the minimum

\item \code{lambda.interp}: Interpolation function.

\item \code{lamfix}: Interpolation function.

\item \code{cvcompute}: Computations for crossvalidation error

\item \code{getmin}: Interpolation function.
}}

\note{
taken verbatim from glmnet

taken verbatim from glmnet
}
\seealso{
\code{\link{cv.sail}}
}
