% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cv.sail}
\alias{cv.sail}
\title{Cross-validation for sail}
\usage{
cv.sail(x, y, e, df, degree, weights, lambda = NULL, type.measure = c("mse",
  "deviance", "class", "auc", "mae"), nfolds = 10, foldid, grouped = TRUE,
  keep = FALSE, parallel = FALSE, ...)
}
\arguments{
\item{x}{Design matrix of dimension \code{n x q}, where \code{n} is the
number of subjects and q is the total number of variables; each row is an
observation vector. This must include all main effects and interactions as
well, with column names corresponding to the names of the main effects
(e.g. \code{x1, x2, E}) and their interactions (e.g. \code{x1:E, x2:E}).
All columns should be scaled to have mean 0 and variance 1; this is done
internally by the \code{\link{sail}} function.}

\item{y}{response variable. For \code{family="gaussian"} should be a 1 column
matrix or numeric vector. For \code{family="binomial"}, if the response is
a vector it can be numeric with 0 for failure and 1 for success, or a
factor with the first level representing "failure" and the second level
representing "success". Alternatively, For binomial logistic regression,
the response can be a matrix where the first column is the number of
"successes" and the second column is the number of "failures".}

\item{weights}{observation weights. Can be total counts if responses are
proportion matrices. Default is 1 for each observation. Currently NOT
IMPLEMENTED}

\item{type.measure}{loss to use for cross-validation. Currently only 1
option. The default is \code{type.measure="mse"}, which uses squared-error
for gaussian models}

\item{nfolds}{number of folds - default is 10. Although nfolds can be as
large as the sample size (leave-one-out CV), it is not recommended for
large datasets. Smallest value allowable is \code{nfolds=3}}

\item{parallel}{If \code{TRUE}, use parallel \code{foreach} to fit each fold.
Must register parallel before hand using the
\code{\link[doMC]{registerDoMC}} function from the \code{doMC} package. See
the example below for details.}
}
\description{
Does k-fold cross-validation for sail and determines the optimal
  pair of tuning parameters (\eqn{\lambda_\beta} and \eqn{\lambda_\gamma})
}
\details{
The function runs sail nfolds+1 times; the first to get the tuning
  parameter sequences, and then the remainder to compute the fit with each of
  the folds omitted. The error is accumulated, and the average error and
  standard deviation over the folds is computed. Note also that the results
  of cv.sail are random, since the folds are selected at random using the
  \code{\link{createfolds}} function. Users can reduce this randomness by
  running cv.sail many times, and averaging the error curves.
}
\author{
Sahir Bhatnagar

  Maintainer: Sahir Bhatnagar \email{sahir.bhatnagar@mail.mcgill.ca}
}
