% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eclust-internal}
\alias{eclust-internal}
\alias{repcol}
\alias{mysd}
\alias{check_col_0}
\alias{nonzero}
\alias{\%ni\%}
\alias{isEmpty}
\alias{checkargs.xy}
\alias{checkargs.misc}
\alias{tabular}
\alias{error.bars}
\title{Internal eclust functions}
\usage{
repcol(x, n)

mysd(i)

check_col_0(M)

nonzero(beta, bystep = FALSE)

... \%ni\% NA

isEmpty(x)

checkargs.xy(x, y)

checkargs.misc(sigma = NULL, alpha = NULL, k = NULL, gridrange = NULL,
  gridpts = NULL, griddepth = NULL, mult = NULL, ntimes = NULL,
  beta = NULL, lambda = NULL, tol.beta = NULL, tol.kkt = NULL,
  bh.q = NULL)

tabular(df, ...)

error.bars(x, upper, lower, width = 0.02, ...)
}
\arguments{
\item{x}{is numeric vector to be repeated}

\item{n}{how many times \code{x} needs to be repeated}

\item{i}{a vector of numerics}

\item{M}{is a matrix}

\item{beta}{vector or 1 column matrix of regression coefficients}
}
\description{
Internal eclust helper functions

\code{repcol} is to return a matrix with \code{x} being repeated
  into \code{n} columns

\code{mysd} is to calculate standard deviation but with divisor of n
  and not n-1

\code{check_col_0} is to check how many columns are 0

\code{nonzero} is to determine which coefficients are non-zero

\code{\%ni\%} is the opposite of \code{\%in\%}

\code{checkargs.xy} function to check inputs of sail function
}
\details{
These functions are not intended for use by users.
}
\author{
Sahir Bhatnagar

  Maintainer: Sahir Bhatnagar \email{sahir.bhatnagar@mail.mcgill.ca}
}
