% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-old.R
\name{gendata}
\alias{gendata}
\title{Simulate Data}
\usage{
gendata(n, p, df, degree, intercept = FALSE, E = rbinom(n = n, size = 1,
  prob = 0.5), betaE = 2, SNR = 1)
}
\description{
function to simulate data
}
