% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotInter}
\alias{plotInter}
\title{Plot Interaction Effects from sail object}
\usage{
plotInter(object, x, xvar, s, f.truth, simulation = TRUE, truthonly = FALSE,
  npoints = 30, col = c("#56B4E9", "#D55E00"), title_z,
  legend.position = "bottomleft", ...)
}
\arguments{
\item{object}{sail object}
}
\description{
Plot Interaction Effects from sail object
}
