% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMain}
\alias{plotMain}
\title{Plot Main Effects from sail object}
\usage{
plotMain(object, xvar, s, f.truth, col = c("#D55E00", "#009E73"),
  legend.position = "bottomleft", ...)
}
\arguments{
\item{object}{sail object}
}
\description{
Plot Main Effects from sail object
}
