% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xtilde}
\alias{xtilde}
\title{Calculate working X's to update Gammas.}
\usage{
xtilde(x, group, main.effect.names.list, interaction.names.list,
  beta.main.effects)
}
\arguments{
\item{x}{data frame or matrix containing all the mains effects, the
environment, and the interaction of the basis expansions}

\item{interaction.names.list}{list of interaction names where each element of
the list corresponds to the original X. Each list element should be of
length df}

\item{beta.main.effects}{matrix containing the coefficients of main effects}

\item{nlambda}{number of tuning parameters}
}
\value{
matrix of working X's (xtilde)
}
\description{
function used to calculate working X's (xtilde) in step 3 of
  algorithm
}
